Added many things)

+++ SummaryTask4/src/ua/nure/zhabin/SummaryTask4/db/dao/UserDao.java
@@ -6,7 +6,7 @@ import ua.nure.zhabin.SummaryTask4.db.entity.User;
 
 public interface UserDao {
 
-	User get(int id, Connection connection);
+	User get(long id, Connection connection);
 	User get(String login, Connection connection);
 	long add(User user, Connection connection);	
 }
